requirements.txt
Django>=5.2,<5.3
wagtail>=7.1,<7.2
django-environ>=0.11.0